+++ lib/features/number_trivia/domain/entities/number_trivia.dart
@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 
 class NumberTriviaEntity extends Equatable {
   final String text;
-  final double number;
+  final int number;
   const NumberTriviaEntity({
     required this.text,
     required this.number,
